Add files via upload
German Windfarm Energy production EDA.

I actually went through this before I completed the Titanic dataset so it's not going to be very good. It would be the very weakest of all my projects. I still think i did a good job with some of the graphs. More work was needed to actually make any sense out of it but as a first attempt I think it was actually okay.